[feature]: add position to noise dataset
An updated events dataset was provided by Roel on 2020-06-09 which was
used instead of the original. Position and direction data was added to
the noise dataset and saved to disk. notebooks/cleanup.py was updated to
reflect all efforts which were made to obtain the final datasets
(noise.csv and hits.csv).