[feature]: init data module
The first of many commits to convert the existing codebase into a
modularized format such that the codebase can be run in a hybrid
manner (python scripts and notebooks) on various devices (locally or
remotely). This commit refactors the code previously used to generate
the noise dataset into the "data" module.