[feature]: refactor data module
Separate methods pertaining to noise, hits and final datasets into
their own modules. This allows for specifying the dataset intended to
be used with the modules and gives a structure to the module. Each
submodule under the data module exposes a top level `process` function
which does the heavy lifting. Alternatively, one may wish to pick and
choose the transformations they wish to perform and create a custom
dataset.