[feature]: add flag to normalize input
Neural Networks perform well when the input to the network is scaler to
a set range (usually dictated by the activation function). A barebone
normalization feature is added to km3net.model.data.CSVDataset where the
input X is normalized between [0, 1] if the normalise flag is set
(defaults to False).