[feature]: support multiple test sets
This commit adds the ability to test a given model with more than one
test dataset. The `km3net.model.utils.evaluate` now expects its
`test_dl` parameter to be a list of DataLoaders. The model is evaluated
iteratively for each test set. The method now returns a list of dicts,
each containing the metrics for each test set in the same order.