[refactor] restructure km3net.model
The pm and gcd models require different methods of training and
testing. At this stage, it is felt apt to restructure the km3net.model
module into two submodules namely pm and gcd which contain the data
prep, the model itself and additional utils for training and testing the
model separately.