[refactor] model data prep from ndarray
Previously, the data prep methods for the models (pm and gcd) expected
a path to a csv from which X, y (and optionally graph edge weights) were
extracted. However, this make the methods rigid and the dev process
cumbersome (as one is expected to first save the csv in the expected
format).

This commit refactors km3net.model.{pm,gcd}.data such that the methods
now require X, y (and z) explicitly passed. This delegates the
responsibility of reading the csv into memory and shaping the data
correctly elsewhere (as it should be).